Move all enums to separate file
